Move bulb code to lib
